Introduce remote cluster security interceptor and authenticator (#134245)
This refactoring is meant to be purely structural, with no functional changes. 
It introduces two interfaces:
- `RemoteClusterAuthenticationService` - extracted based on the existing `CrossClusterAccessAuthenticationService`
- `RemoteClusterTransportInterceptor` - which aims to abstract and move all "cross-cluster access" logic from `SecurityServerTransportInterceptor` 

This is prerequisite for making remote cluster security logic pluggable (which I plan to add in a followup PR).

Relates to ES-12801
