+++ x-pack/plugin/security/src/main/java/org/elasticsearch/xpack/security/authz/PreAuthorizationUtils.java
@@ -23,7 +23,6 @@ import org.elasticsearch.xpack.core.security.authz.permission.Role;
 
 import java.util.Arrays;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
 
 public final class PreAuthorizationUtils {
@@ -118,9 +117,9 @@ public final class PreAuthorizationUtils {
     }
 
     public static boolean shouldRemoveParentAuthorizationFromThreadContext(
-        Optional<String> remoteClusterAlias,
         String childAction,
-        SecurityContext securityContext
+        SecurityContext securityContext,
+        boolean isRemoteClusterRequest
     ) {
         final ParentActionAuthorization parentAuthorization = securityContext.getParentAuthorization();
         if (parentAuthorization == null) {
@@ -128,7 +127,7 @@ public final class PreAuthorizationUtils {
             return false;
         }
 
-        if (remoteClusterAlias.isPresent()) {
+        if (isRemoteClusterRequest) {
             // We never want to send the parent authorization header to remote clusters.
             return true;
         }
